requirements.txt
google-generativeai
python-dotenv
flask
fastapi
uvicorn
flask-cors
requests